+++ src/main/java/io/github/astrapi69/io/file/FileExtension.java
@@ -24,6 +24,9 @@
  */
 package io.github.astrapi69.io.file;
 
+import java.io.File;
+import java.util.Objects;
+
 /**
  * The enum {@link FileExtension} for file extensions
  */
@@ -101,6 +104,40 @@ public enum FileExtension
 	public String getExtension()
 	{
 		return extension;
+	}
+
+	/**
+	 * Checks if the given file is from the given {@link FileExtension} object
+	 * 
+	 * @param file
+	 *            the file to check
+	 * @param fileExtension
+	 *            the file extension
+	 * @return true if the given file is from the given {@link FileExtension} object otherwise false
+	 */
+	public static boolean is(File file, FileExtension fileExtension)
+	{
+		Objects.requireNonNull(file, "File must not be null");
+		Objects.requireNonNull(file, "FileExtension must not be null");
+		return is(file, fileExtension.getExtension());
+	}
+
+	/**
+	 * Checks if the given file is from the given file extension as {@link String} object
+	 * 
+	 * @param file
+	 *            the file to check
+	 * @param fileExtension
+	 *            the file extension as {@link String} object
+	 * @return true if the given file is from the given {@link String} object otherwise false
+	 */
+	public static boolean is(File file, String fileExtension)
+	{
+		Objects.requireNonNull(file, "File must not be null");
+		Objects.requireNonNull(file, "String for the file extension must not be null");
+		return fileExtension.startsWith(".")
+			? FilenameExtensions.getFileExtension(file, true).equals(fileExtension)
+			: FilenameExtensions.getFileExtension(file, false).equals(fileExtension);
 	}
 
 }